docs: migrated contract methods
