+++ apps/docs/src/guide/contracts/methods.md
@@ -6,13 +6,13 @@ There are 4 ways to interact with contracts: `get`, `dryRun`, `simulate`, `call`
 
 The `get` method should be used to read data from the blockchain without using resources. It can be used with an unfunded wallet or even without a wallet at all:
 
-<<< @/../../docs-snippets/src/guide/contracts/interacting-with-contracts.test.ts#interacting-with-contracts-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/methods/get.ts#interacting-with-contracts-1{ts:line-numbers}
 
 ## `dryRun`
 
 The `dryRun` method should be used to dry-run a contract call. It does not spend resources and can be used with an unfunded wallet or even without a wallet at all:
 
-<<< @/../../docs-snippets/src/guide/contracts/interacting-with-contracts.test.ts#interacting-with-contracts-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/methods/dry-run.ts#interacting-with-contracts-2{ts:line-numbers}
 
 ## `simulate`
 
@@ -20,7 +20,7 @@ The `simulate` method should be used to dry-run a contract call, ensuring that t
 
 A funded wallet it's required:
 
-<<< @/../../docs-snippets/src/guide/contracts/interacting-with-contracts.test.ts#interacting-with-contracts-3{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/methods/simulate.ts#interacting-with-contracts-3{ts:line-numbers}
 
 ## `call`
 
@@ -28,13 +28,13 @@ The `call` method submits a real contract call transaction to the node, resolvin
 
 Real resources are consumed, and any operations executed by the contract function will be processed on the blockchain.
 
-<<< @/../../docs-snippets/src/guide/contracts/interacting-with-contracts.test.ts#interacting-with-contracts-4{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/methods/call.ts#interacting-with-contracts-4{ts:line-numbers}
 
 ## `isReadOnly` (utility)
 
 If you want to figure out whether a function is read-only, you can use the `isReadOnly` method:
 
-<<< @/../../docs-snippets/src/guide/contracts/is-function-readonly.test.ts#is-function-readonly-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/methods/is-read-only.ts#is-function-readonly-1{ts:line-numbers}
 
 If the function is read-only, you can use the `get` method to retrieve onchain data without spending gas.
